refactor: Remove user field from Participant model and replace with email field
The user field in the Participant model has been removed and replaced with an email field. This change was made to decouple the Participant model from the authentication system and allow for more flexibility in handling participant data. The email field is now used as the identifier for participants. This commit addresses the need to separate the user authentication from the participant data in the system.