refactor: Update TransactionCreateView to include reference field in response
The TransactionCreateView in the payments/views.py file has been updated to include the "reference" field in the response data. This change allows for the successful creation of a transaction to include the reference value. The reference field is a string that uniquely identifies the transaction. This commit addresses the need to enhance the functionality of the TransactionCreateView by incorporating the reference field in the response.