refactor: Include reference field in RegisterView response
The RegisterView in the register/views.py file has been updated to include the "reference" field in the response data. This change allows for the successful registration of a participant to include the reference value. The reference field is a string that uniquely identifies the participant. This commit addresses the need to enhance the functionality of the RegisterView by incorporating the reference field in the response.