feat: Include reference field in GenerateSerializer and GeneratePaymentLinkAPIView
The `GenerateSerializer` class in the `register/serializers.py` file has been updated to include a `reference` field. This change allows for the serialization of the `reference` attribute in the API response.

The `GeneratePaymentLinkAPIView` in the `register/views.py` file has been modified to include the `reference` field in the response. This change provides more detailed information in the API response.